mcma: 2nd approx. of nadir finished.
Also, in PWL ctor diagnostics of criteria attributes (U, A, R, N) and their relations is now implemented. Diverse improvements in control of analysis flow. Remaining issues are in todo. More testing is needed but the app is ready for implementation of the last main stage, i.e., specification and handling the preferences (A, R, crit-activity).