mcma: tested pipa model analysis.
Generation of the (old) pipa model, storing and reading it from *.dll file proved that the intended modularity works (at least on the macOs; portability of *dll format between macOS and other OSs needs to be explored). However, the MCMA of the old version of the pipa model does not produce "interesting" results (criteria defined by the outcome variable appears to be linearly depended). Therefore, for further mcma development and testing the (to be improved) sbPipa model will be used until a new pipa model will be available).